pyproject.toml
[project]
name = "bioagent-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "smolagents[docker,openai]>=1.21.2",
    "docker>=7.1.0",
    "openinference-instrumentation-smolagents>=0.1.16",
    "arize-phoenix>=11.30.0",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
bioagent-experiments = { workspace = true }

[dependency-groups]
dev = [
    "bioagent-experiments",
]

[tool.setuptools.packages.find]
where = ["src"]
